requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-oauth-toolkit==3.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography==42.0.5
